Reorganiza conteúdo de basenotes em basenotes, fn e author_notes
